+++ src/utils/getApiKey.ts
@@ -1,11 +1,11 @@
 import * as vscode from "vscode";
 
 export function getApiKey() {
-  const config = vscode.workspace.getConfiguration("askDevAI");
+  const config = vscode.workspace.getConfiguration("askDevAIchatbot");
 
   if (config.apiKey.length <= 0) {
     vscode.commands.executeCommand(
-      "DevAI.showErrorNotification",
+      "DevAIchatbot.showErrorNotification",
       "Please, set up your api key"
     );
     return "";