+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the "DevAI" extension will be documented in this file.
+All notable changes to the "DevAIchatbot" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
@@ -11,4 +11,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### 0.0.2
 
 - Create setting to add your OpenAI api key
-- AskDevAI feature is fully functional
+- AskDevAIchatbot feature is fully functional